Initial import of the Dump module from Pyromaniac.
The RISC OS Pyromaniac module for dumping content can be made accessible
for other users. It's not especially complicated, but it does a good
job. It needs a bit more explanation and some more functional changes
to be more generally useful, but this separate part should be able to
be used by others.