Improve performance and make resize work better.
The grid was using AutoResize which attempts to read every single
cell in the grid in order to get the correct sizes. This was happening
in column-major ordering, which defeated out row cache. The result was
that larger documents were incredibly slow to render.

Resizes are now using the correct sizes of text so that the window isn't
too small and the scrollbar appears correctly.