Disable the selection of non-data cells.
The data cells should be the only cells that can be selected. This
means that the cursor will only ever be in the cells that represent
the data, and not the interpretation cells that are the text and
annotations.

This change prevents the cursor moving our of the cells that contain
the data.