Fix wxdump.py for small windows and refreshing data not working.
Small windows might give us a backtrace because the minimum size was
larger than the natural size. This was because in small windows we would
only size to the data that was present and thus not show things right.
We now enforce a minimum on the size.

If you needed to refresh the data in the dump window this would need
the row cache clearing. This wasn't obvious and so in the RISC OS
Pyromaniac windows it wasn't working. We now have a SetData function
to do what's needed to update the content.